python_app/requirements.txt
flask
psycopg2-binary
requests

# OpenTelemetry
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-psycopg2

# Prometheus
prometheus_flask_exporter
